+++ app/src/main/java/hypr/a255bits/com/hypr/GeneratorLoader/FaceLocation.kt
@@ -1,0 +1,7 @@
+package hypr.a255bits.com.hypr.GeneratorLoader
+
+import android.graphics.Bitmap
+import android.graphics.Rect
+
+
+class FaceLocation(val croppedFace: Bitmap, val faceLocation: Rect)
